Fusion get a (bigger) update
When Fusion get a (bigger) update then it could happen that a new folder is created in which the Fusion360.exe is also located!
With this change, the latest version will be selected and started up for you. :+1:

Thanks to the user @Tony763 for his [help](https://github.com/cryinkfly/Fusion-360---Linux-Wine-Version-/issues/71#issue-982958482)!