Change invocation of mokutil to return 0 on success
Some versions of mokutil have an exit code of -1 when invoked without
arguments, which causes the check for mokutil to fail and subsequent
re-installation of required tools, regardless of whether they are
already installed or not.
Invoking mokutil with --version also results in an exit code of -1. This
seems to have been fixed in (at least) mokutil 0.7.0.